check if files exist before copy

+++ install.sh
@@ -1,9 +1,18 @@
-mv ~/.bashrc ~/.bashrc.bak
-mv ~/.tmux.conf ~/.tmux.conf.bak
-mv ~/vimrc ~/vimrc.bak
+if [ -f ~/.bashrc ];then
+  mv ~/.bashrc ~/.bashrc.bak
+fi
+if [ -f ~/.tmux.conf ];then
+  mv ~/.tmux.conf ~/.tmux.conf.bak
+fi
+if [ -f ~/.vimrc ];then
+  mv ~/.vimrc ~/.vimrc.bak
+fi
+if [ -f ~/.gitconfig ];then
+  mv ~/.gitconfig ~/.gitconfig.bak
+fi
 
-ln -sc "~/dotfiles/.bashrc" ~
-ln -sc "~/dotfiles/.tmux.conf" ~
-ln -sc "~/dotfiles/vimrc" ~
-ln -sc "~/dotfiles/.gitconfig" ~
+ln -s ~/dotfiles/.bashrc ~/.bashrc
+ln -s ~/dotfiles/.tmux.conf ~/.tmux.conf
+ln -s ~/dotfiles/.vimrc ~/.vimrc
+ln -s ~/dotfiles/.gitconfig ~/.gitconfig
